requirements.txt
asgiref==3.3.0
Django==3.0.10
django-filter==2.4.0
joblib==0.17.0
numpy==1.19.3
pandas
Pillow
python-dateutil==2.8.1
pytz==2020.1
scikit-learn==0.23.2
scipy
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
agora
agora-python-sdk
django-cors-headers
requests